Added Authorization by Reservation (not tested, work in progress)

+++ FoodFusion/Controllers/ReservationsController.cs
@@ -1,11 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Security.Claims;
-using System.Threading.Tasks;
 using Common;
-using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Services.Authorization;
 using Services.Authorization.Exceptions;
 using Services.Reservations;
 using Services.Reservations.Exceptions;
@@ -18,18 +16,22 @@ namespace WebApi.Controllers
     [ApiController]
     public class ReservationsController : ControllerBase
     {
+        private readonly IResourceAuthorizationService<ReservationAuthorizationRequirement> _reservationAuthorizationService;
         private readonly IReservationsService _reservationsService;
         private readonly IAvailabilityService _availabilityService;
 
         public ReservationsController(
+            IResourceAuthorizationService<ReservationAuthorizationRequirement> reservationAuthorizationService,
             IReservationsService reservationsService, 
             IAvailabilityService availabilityService)
         {
+            _reservationAuthorizationService = reservationAuthorizationService;
             _reservationsService = reservationsService;
             _availabilityService = availabilityService;
         }
 
         // GET: api/Restaurants/5/unavailability
+        [Authorize]
         [HttpGet("Restaurants/{restaurantId}/Unavailability")]
         public IActionResult GetUnavailability(
             [FromRoute] int restaurantId, 
@@ -46,6 +48,7 @@ namespace WebApi.Controllers
         }
 
         // POST: api/Restaurant/5/Reservations
+        [Authorize]
         [HttpPost("Restaurants/{restaurantId}/Reservations")]
         public IActionResult Post([FromRoute] int restaurantId, [FromBody] ReservationRequestModel reservationRequest)
         {
@@ -62,6 +65,7 @@ namespace WebApi.Controllers
         }
 
         // GET: api/Restaurant/5/Reservations
+        [Authorize]
         [HttpGet("Restaurant/{restaurantId}/Reservations")]
         public IActionResult GetRestaurantReservations(int restaurantId)
         {
@@ -71,6 +75,7 @@ namespace WebApi.Controllers
         }
 
         // GET: api/Reservations
+        [Authorize]
         [HttpGet("Reservations")]
         public IActionResult GetUserReservations()
         {
@@ -83,9 +88,17 @@ namespace WebApi.Controllers
         }
 
         // GET: api/Reservations/5
+        [Authorize]
         [HttpGet("Reservations/{id}")]
         public IActionResult Get(int id)
         {
+            var isAuthorized = _reservationAuthorizationService
+                .WithUser(User)
+                .WithRequirement(Operations<ReservationAuthorizationRequirement>.Read)
+                .WithResource(id)
+                .IsAuthorized();
+            if (!isAuthorized) return Forbid();
+
             try
             {
                 var reservation = _reservationsService.GetReservation(id);
@@ -99,9 +112,17 @@ namespace WebApi.Controllers
         }
 
         // PUT: api/Restaurant/5/Reservations
+        [Authorize]
         [HttpPut("Reservations/{id}")]
         public IActionResult Put(int id, [FromBody] ReservationRequestModel reservationRequest)
         {
+            var isAuthorized = _reservationAuthorizationService
+                .WithUser(User)
+                .WithRequirement(Operations<ReservationAuthorizationRequirement>.Update)
+                .WithResource(id)
+                .IsAuthorized();
+            if (!isAuthorized) return Forbid();
+
             if (!ModelState.IsValid) return BadRequest(ModelState);
 
             reservationRequest.Id = id;
@@ -118,9 +139,17 @@ namespace WebApi.Controllers
         }
 
         // DELETE: api/Reservations/5
+        [Authorize]
         [HttpDelete("Reservations/{id}")]
         public IActionResult Delete(int id)
         {
+            var isAuthorized = _reservationAuthorizationService
+                .WithUser(User)
+                .WithRequirement(Operations<ReservationAuthorizationRequirement>.Delete)
+                .WithResource(id)
+                .IsAuthorized();
+            if (!isAuthorized) return Forbid();
+
             try
             {
                 _reservationsService.RemoveReservation(id);
